requirements.txt
robotframework
robotframework-seleniumlibrary
selenium==4.2.0
six
